# openapi.yaml
openapi: 3.0.3
info:
  title: Authly – Authentification Universelle
  description: >
    Service SaaS pour gérer l'authentification OAuth2, la gestion des rôles, 
    la sécurité via JWT et l'intégration de Firebase/GCP.  
    Cette specification OpenAPI décrit l'API de Authly – Authentification Universelle.
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:8000
    description: Serveur local de développement
  - url: https://api.authly.io
    description: Serveur de production (exemple)
...
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Authentifier un utilisateur
      description: >
        Authentifie un utilisateur via email/mot de passe **ou** via un code OAuth 
        obtenu d'un provider (Google, GitHub, Facebook).  
        Retourne un JWT (ou un token Firebase) si la connexion est validée.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Succès de la connexion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Données invalides (format, champs manquants, etc.)
        '401':
          description: Échec de l'authentification (identifiants incorrects)
        '500':
          description: Erreur interne du serveur
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Déconnecter un utilisateur
      description: >
        Invalide la session ou le token de l'utilisateur connecté.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Succès de la déconnexion
        '401':
          description: Token invalide ou expiré
        '500':
          description: Erreur interne du serveur
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Rafraîchir un token
      description: >
        Permet de rafraîchir le token JWT ou d'obtenir un nouveau jeton 
        si l'ancien est expiré mais que l'utilisateur dispose d'un refresh_token valide.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Succès de l'opération, renvoie un nouveau token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Jeton de rafraîchissement invalide ou expiré
        '500':
          description: Erreur interne
  /auth/oauth/{provider}:
    get:
      tags:
        - Auth
      summary: Déclencher un flux OAuth avec un provider
      description: >
        Redirige vers le provider OAuth spécifié (google, github, facebook).  
        L'URL de callback, côté front-end, traitera la réponse pour ensuite appeler `/auth/login`.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, facebook]
          description: Provider OAuth à utiliser
      responses:
        '302':
          description: Redirection vers la page d'authentification du provider
        '400':
          description: Provider non supporté
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        oauth_code:
          type: string
          description: Code OAuth renvoyé par un provider (facultatif)
        provider:
          type: string
          description: Provider OAuth (google, github, facebook)
      example:
        email: "johndoe@example.com"
        password: "mypassword"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: Jeton JWT ou token Firebase
        token_type:
          type: string
          description: Type du jeton
        expires_in:
          type: integer
          description: Durée de validité du jeton en secondes
        refresh_token:
          type: string
          description: Token de rafraîchissement
      example:
        access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type: "Bearer"
        expires_in: 3600
        refresh_token: "Rfresht0ken..."
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
      example:
        access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type: "Bearer"
        expires_in: 3600

    # ========================
    # Users
    # ========================
...